Restore Rails 2 support
Wahen include ActionController::Base.include(::SecureHeaders) called from reiltie.rb,
SecuerHeaders.add_features not defined yet.

As a result class/instance methods won't added to ActionController::Base.

But introduced in commit 4aee15ef85c7413aa6718a1ec9ea82f843eb6d03
